--- app/src/commonMain/kotlin/com/maksimowiczm/foodyou/core/data/openfoodfacts/model/v1/OpenFoodFactsProductV1.kt
@@ -1,26 +1,0 @@
-package com.maksimowiczm.foodyou.core.data.openfoodfacts.model.v1
-
-import com.maksimowiczm.foodyou.core.domain.model.openfoodfacts.OpenFoodFactsNutrients
-import com.maksimowiczm.foodyou.core.domain.model.openfoodfacts.OpenFoodFactsProduct
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class OpenFoodFactsProductV1(
-    @SerialName("product_name")
-    override val productName: String? = null,
-    @SerialName("brands")
-    override val brands: String? = null,
-    @SerialName("code")
-    override val code: String? = null,
-    @SerialName("nutriments")
-    override val nutrients: OpenFoodFactsNutrients? = null,
-    @SerialName("product_quantity")
-    override val packageQuantity: Float? = null,
-    @SerialName("product_quantity_unit")
-    override val packageQuantityUnit: String? = null,
-    @SerialName("serving_quantity")
-    override val servingQuantity: Float? = null,
-    @SerialName("serving_quantity_unit")
-    override val servingQuantityUnit: String? = null
-) : OpenFoodFactsProduct
