+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/data/ProductNotFoundException.kt
@@ -1,0 +1,3 @@
+package com.maksimowiczm.foodyou.feature.product.data
+
+internal class ProductNotFoundException : Exception()
