+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/data/network/openfoodfacts/model/v2/OpenFoodFactsProductResponseV2.kt
@@ -1,4 +1,4 @@
-package com.maksimowiczm.foodyou.feature.productdownload.data.network.openfoodfacts.model.v2
+package com.maksimowiczm.foodyou.feature.product.data.network.openfoodfacts.model.v2
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
