+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/domain/RemoteProduct.kt
@@ -1,0 +1,5 @@
+package com.maksimowiczm.foodyou.feature.product.domain
+
+interface RemoteProduct {
+    val name: String?
+}
