+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/domain/RemoteProductMatcher.kt
@@ -1,4 +1,4 @@
-package com.maksimowiczm.foodyou.feature.productdownload.domain
+package com.maksimowiczm.foodyou.feature.product.domain
 
 fun interface RemoteProductMatcher {
     fun matches(url: String): Boolean
