--- app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/productdownload/data/ProductNotFoundException.kt
@@ -1,3 +1,0 @@
-package com.maksimowiczm.foodyou.feature.productdownload.data
-
-internal class ProductNotFoundException : Exception()
