--- app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/productdownload/domain/RemoteProduct.kt
@@ -1,5 +1,0 @@
-package com.maksimowiczm.foodyou.feature.productdownload.domain
-
-interface RemoteProduct {
-    val name: String?
-}
