+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/infrastructure/di/InitKoin.kt
@@ -22,7 +22,6 @@ fun initKoin(config: KoinAppDeclaration? = null) = startKoin {
         dataStoreModule,
         domainModule,
         networkModule,
-        openFoodFactsModule,
         utilsModule
     )
